openleaf-reader-platform/services/publishing-service/requirements.txt
flask
pymongo
pika
pyjwt
